requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
apscheduler==3.10.4
python-dotenv==1.0.1
